task_4/scaffolddjango/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-model-utils
